Moved ScheduledJDBCMetricsCleanupTask to org.wso2.carbon.metrics.impl component
